Chore/fe/lint fix (#259)
* admin dashboard waiting number fix

* dialog layout modify

* publisher refator

* publisher stats tab upgrade

* admin page, publisher page routing

* alert -> snackbar
