+++ client/src/api/api.ts
@@ -10,7 +10,6 @@ const api = new Api<undefined>({
   },
   securityWorker: async () => {
     const user = getDefaultStore().get(State.Auth.user);
-    console.log(user);
     if (!user) {
       return {};
     }
