+++ client/src/component/CreditPurchaseModal.tsx
@@ -14,7 +14,7 @@ import { useSnackbar } from "notistack";
 import API_CLIENT from "../api/api";
 import { useEffect, useState } from "react";
 import { paymentSuccessToken } from "../types/paymentSuccessMessage";
-import { setResponsiveStyleValueSm } from "../util/setResponsiveStyleValue";
+import { setResponsiveStyleValueSm } from "../utils/setResponsiveStyleValue";
 
 export default function CreditPurchaseModal(props: {
   open: boolean;
