+++ client/src/component/tabs/TabPanel.tsx
@@ -1,0 +1,20 @@
+import React from "react";
+import { Box } from "@mui/material";
+
+interface TabPanelProps {
+  children?: React.ReactNode;
+  index: number;
+  value: number;
+}
+
+export const TabPanel: React.FC<TabPanelProps> = ({
+  children,
+  value,
+  index,
+}) => {
+  return (
+    <div hidden={value !== index}>
+      {value === index && <Box sx={{ py: 3 }}>{children}</Box>}
+    </div>
+  );
+};
