+++ client/src/routes/__root.tsx
@@ -12,6 +12,7 @@ import {
   BookOnline,
   Group,
   Home,
+  ImportContacts,
 } from "@mui/icons-material";
 import { useAtomValue } from "jotai";
 import State from "../states";
@@ -35,9 +36,13 @@ const defaultNavigationItems: NavigationItem[][] = [
 ];
 
 const adminNavigationItems: NavigationItem[] = [
-  { to: "/admin", label: "관리자", icon: <AdminPanelSettings /> },
+  { to: "/mypage/admin", label: "관리자", icon: <AdminPanelSettings /> },
   { to: "/admin/books", label: "도서 관리", icon: <BookOnline /> },
   { to: "/admin/publisher", label: "출판사 관리", icon: <Badge /> },
+];
+
+const publisherNavigationItems: NavigationItem[] = [
+  { to: "/mypage/publisher", label: "출판사", icon: <ImportContacts /> },
 ];
 
 autoLogin();
@@ -52,6 +57,7 @@ function RouteComponent() {
   );
 
   const isAdmin = user && user.role === Role.ROLE_ADMIN;
+  const isPublisher = user && user.role === Role.ROLE_PUBLISHER;
 
   return (
     <>
@@ -61,6 +67,7 @@ function RouteComponent() {
             sideNavigationBarItemsWithGroups={[
               ...defaultNavigationItems,
               ...(isAdmin ? [adminNavigationItems] : []),
+              ...(isPublisher ? [publisherNavigationItems] : []),
             ]}
           />
           <Toolbar />
