+++ client/src/utils/loadLocations.ts
